move fragments into separated folder
