+++ tests/page_objects/browse_issue_page_object.py
@@ -1,6 +1,6 @@
 from tests.page_objects.base_page_object import BasePageObject
-from tests.page_objects.header_toolbar_fragment import HeaderToolbarFragment
-from tests.page_objects.issue_details_fragment import IssueDetailsFragment
+from tests.page_objects.fragments.header_toolbar_fragment import HeaderToolbarFragment
+from tests.page_objects.fragments.issue_details_fragment import IssueDetailsFragment
 
 
 class BrowseIssuePageObject(BasePageObject):
