+++ tests/page_objects/fragments/issue_details_fragment.py
